Moved the function to create table body of refile error to an independant file.
